Add site wide and coins page specific loading indicators
Now when the user navigates throughout the app, there will be a progress
bar loading indicator present right beneath the top header.

Also when the user navigates to the coins component, there will be a
spinner that is shown until all the data that needs to be displayed is
ready to be shown.