Fix HomePage and Coins Component loading issues
HomePage Component was not loading. This was due to the ngOnInit logic not
being updated to reflect the new ngOnInit logic added to the
CoinsComponent. Specifically the need to assign an empty object to each
coin element's SocketData property, and the need to prepopulate BTC and or
ETH data if the conversion currency in question was BTC or ETH
respectively.

Coins Component was loading too quickly (before data had arrived). Changed
logic within ngOnInit such that loading will be set to false only if
socket data has actually been recieved.