Fix bug where sometimes data would not load on HomePage Component
This was due to Crypcompare socket volume data still being emitted when
navigating from the CoinsComponent to the HomePage Component. This hit a
corner case in the CoinsService's method that handles incoming
Cryptocompare volume socket data, that threw an exception.

With this fix, there is now support within the CryptocompareSocketService
to close and re-open a socket connection. This ability to close the socket
is what ultimately rectifies the issue at hand, as previously the socket
was never closed, but instead only had subscriptions removed when
navigating from CoinsComponent to HomePage Component.